refactor - Deserialization 위한 라이브러리 변경 후 적용
