+++ app/build.gradle
@@ -2,6 +2,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlinx-serialization'
 
 android {
     compileSdkVersion 30
@@ -68,8 +69,6 @@ dependencies {
 
     implementation "com.kakao.sdk:v2-user:2.0.3" // 카카오 로그인
 
-    implementation 'com.beust:klaxon:5.0.1' // json 변환
-
     implementation 'com.github.bishoybasily:stomp:2.0.5' // stomp
     implementation 'com.squareup.okhttp3:okhttp:4.3.1'
 
@@ -77,5 +76,12 @@ dependencies {
 
     implementation "com.github.bumptech.glide:glide:$glide_version"
     annotationProcessor "com.github.bumptech.glide:compiler:$glide_version"
+
+    // JsonObject 변환
+    implementation 'com.beust:klaxon:5.0.1'
+
+    // kotlin serialization
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.20.0" // JVM dependency
 
 }
