+++ build.gradle
@@ -17,6 +17,8 @@ buildscript {
 
     }
     dependencies {
+
+        classpath "org.jetbrains.kotlin:kotlin-serialization:$kotlin_version"
         classpath "com.android.tools.build:gradle:4.0.1"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
